backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.5
apscheduler==3.10.4
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
authlib==1.2.1
itsdangerous==2.1.2
email-validator==2.1.0.post1